feat: improve validation errors
