+++ lib/Server.js
@@ -57,7 +57,10 @@ class Server {
       throw new Error("'filename' option must be set in lazy mode.");
     }
 
-    validateOptions(schema, options, 'webpack Dev Server');
+    validateOptions(schema, options, {
+      name: 'Dev Server',
+      baseDataPath: 'options',
+    });
 
     this.compiler = compiler;
     this.options = options;
