+++ package.json
@@ -55,9 +55,9 @@
     "open": "^7.0.0",
     "p-retry": "^4.1.0",
     "portfinder": "^1.0.25",
-    "schema-utils": "^1.0.0",
+    "schema-utils": "^2.6.4",
     "selfsigned": "^1.10.7",
-    "semver": "^7.1.0",
+    "semver": "^7.1.1",
     "serve-index": "^1.9.1",
     "sockjs": "0.3.19",
     "sockjs-client": "1.4.0",
@@ -82,16 +82,16 @@
     "body-parser": "^1.19.0",
     "commitlint-azure-pipelines-cli": "^1.0.3",
     "copy-webpack-plugin": "^5.1.1",
-    "css-loader": "^2.1.1",
+    "css-loader": "^3.4.2",
     "eslint": "^6.8.0",
     "eslint-config-prettier": "^6.10.0",
     "eslint-config-webpack": "^1.2.5",
     "eslint-plugin-import": "^2.20.0",
-    "execa": "^4.0.3",
+    "execa": "^4.0.0",
     "file-loader": "^5.0.2",
     "html-loader": "^0.5.5",
     "html-webpack-plugin": "^3.2.0",
-    "husky": "^3.1.0",
+    "husky": "^4.2.1",
     "jest": "^24.9.0",
     "jest-junit": "^10.0.0",
     "jquery": "^3.4.1",
